requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
python-socketio==5.11.3
python-engineio==4.9.1
eventlet==0.36.1
Werkzeug==3.0.3
